[wallet/symbol/mobile] task: move RootLayout into AppRootTemplate and its inner components
