+++ wallet/symbol/mobile/src/components/layout/index.js
@@ -6,13 +6,9 @@ export * from './Field';
 export * from './FlexContainer';
 export * from './Grid';
 export * from './LoadingIndicator';
-export * from './NetworkConnectionStatusBar';
 export * from './Passcode';
-export * from './PopupMessage';
-export * from './RootLayout';
 export * from './Screen';
 export * from './Spacer';
 export * from './Stack';
 export * from './StatusCard';
 export * from './StyledText';
-export * from './SystemStatusBar';
