+++ wallet/symbol/mobile/src/components/templates/AppRootTemplate/components/NetworkConnectionStatusBar.jsx
