+++ wallet/symbol/mobile/src/components/templates/AppRootTemplate/components/SystemStatusBar.jsx
