+++ wallet/symbol/mobile/src/components/templates/AppRootTemplate/components/index.js
@@ -1,0 +1,3 @@
+export { NetworkConnectionStatusBar } from './NetworkConnectionStatusBar';
+export { PopupMessage } from './PopupMessage';
+export { SystemStatusBar } from './SystemStatusBar';
