+++ wallet/symbol/mobile/src/components/templates/index.js
@@ -1,0 +1,2 @@
+export { AppRootTemplate } from './AppRootTemplate/AppRootTemplate';
+export { TransactionScreenTemplate } from './TransactionScreenTemplate/TransactionScreenTemplate';
